# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Football Kickoff 2026 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments/{slug}/matches:
    get:
      operationId: listMatches
      tags: [matches]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
        - in: query
          name: status
          required: false
          schema: { type: string, enum: [open, locked, completed, all] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MatchWithPrediction" }
    post:
      operationId: createMatch
      tags: [matches, admin]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MatchInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Match" }
  /tournaments/{slug}/matches/import:
    post:
      operationId: importMatchesCsv
      tags: [matches, admin]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CsvImportInput" }
      responses:
        "200":
          description: Imported
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportSummary" }
  /matches/{id}:
    patch:
      operationId: updateMatch
      tags: [matches, admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MatchUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Match" }
    delete:
      operationId: deleteMatch
      tags: [matches, admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /matches/{id}/result:
    post:
      operationId: enterMatchResult
      tags: [matches, admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MatchResultInput" }
      responses:
        "200":
          description: Result saved and predictions scored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Match" }
  /matches/{id}/recalculate:
    post:
      operationId: recalculateMatch
      tags: [matches, admin]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Recalculated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Match" }
  /matches/{id}/predict:
    post:
      operationId: submitPrediction
      tags: [predictions]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PredictionInput" }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Prediction" }
...
  /matches/{id}/lineups:
    get:
      operationId: getMatchLineups
      tags: [public]
      summary: Starting lineups + formations for both teams in a match
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MatchLineups" }
        "404":
          description: Match not found
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, name, code, flag]
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        flag: { type: string }
        primaryColor: { type: ["string", "null"] }
        secondaryColor: { type: ["string", "null"] }
        crestUrl: { type: ["string", "null"] }
...
    Match:
      type: object
      required: [id, tournamentId, round, group, teamA, teamB, kickoffTime, lockTime, status]
      properties:
        id: { type: integer }
        tournamentId: { type: integer }
        round: { type: string }
        group: { type: string }
        teamA: { $ref: "#/components/schemas/Team" }
        teamB: { $ref: "#/components/schemas/Team" }
        kickoffTime: { type: string, format: date-time }
        lockTime: { type: string, format: date-time }
        status: { type: string, enum: [open, locked, completed] }
        scoreA: { type: ["integer", "null"] }
        scoreB: { type: ["integer", "null"] }
        extraTimeScoreA: { type: ["integer", "null"] }
        extraTimeScoreB: { type: ["integer", "null"] }
        penaltiesScoreA: { type: ["integer", "null"] }
        penaltiesScoreB: { type: ["integer", "null"] }
        duration:
          type: ["string", "null"]
          enum: ["REGULAR", "EXTRA_TIME", "PENALTY_SHOOTOUT", null]
    MatchWithPrediction:
      type: object
      required: [id, tournamentId, round, group, teamA, teamB, kickoffTime, lockTime, status, isLocked]
      properties:
        id: { type: integer }
        tournamentId: { type: integer }
        round: { type: string }
        group: { type: string }
        teamA: { $ref: "#/components/schemas/Team" }
        teamB: { $ref: "#/components/schemas/Team" }
        kickoffTime: { type: string, format: date-time }
        lockTime: { type: string, format: date-time }
        status: { type: string, enum: [open, locked, completed] }
        scoreA: { type: ["integer", "null"] }
        scoreB: { type: ["integer", "null"] }
        extraTimeScoreA: { type: ["integer", "null"] }
        extraTimeScoreB: { type: ["integer", "null"] }
        penaltiesScoreA: { type: ["integer", "null"] }
        penaltiesScoreB: { type: ["integer", "null"] }
        duration:
          type: ["string", "null"]
          enum: ["REGULAR", "EXTRA_TIME", "PENALTY_SHOOTOUT", null]
        isLocked: { type: boolean }
        myPrediction:
          oneOf:
            - { $ref: "#/components/schemas/Prediction" }
            - { type: "null" }
    MatchInput:
      type: object
      required: [teamAId, teamBId, kickoffTime, lockTime]
      properties:
        round: { type: string }
        group: { type: string }
        teamAId: { type: integer }
        teamBId: { type: integer }
        kickoffTime: { type: string, format: date-time }
        lockTime: { type: string, format: date-time }
        status: { type: string, enum: [open, locked, completed] }
    MatchUpdate:
      type: object
      properties:
        round: { type: string }
        group: { type: string }
        teamAId: { type: integer }
        teamBId: { type: integer }
        kickoffTime: { type: string, format: date-time }
        lockTime: { type: string, format: date-time }
        status: { type: string, enum: [open, locked, completed] }
        scoreA: { type: integer }
        scoreB: { type: integer }
    MatchResultInput:
      type: object
      required: [scoreA, scoreB]
      properties:
        scoreA: { type: integer, minimum: 0 }
        scoreB: { type: integer, minimum: 0 }
    PredictionInput:
      type: object
      required: [predictedScoreA, predictedScoreB]
      properties:
        predictedScoreA: { type: integer, minimum: 0, maximum: 99 }
        predictedScoreB: { type: integer, minimum: 0, maximum: 99 }
        predictedExtraTimeA: { type: ["integer", "null"], minimum: 0, maximum: 99 }
        predictedExtraTimeB: { type: ["integer", "null"], minimum: 0, maximum: 99 }
        predictedPenaltiesA: { type: ["integer", "null"], minimum: 0, maximum: 30 }
        predictedPenaltiesB: { type: ["integer", "null"], minimum: 0, maximum: 30 }
    Prediction:
      type: object
      required: [id, tournamentId, matchId, userId, predictedScoreA, predictedScoreB, status, submittedAt]
      properties:
        id: { type: integer }
        tournamentId: { type: integer }
        matchId: { type: integer }
        userId: { type: integer }
        predictedScoreA: { type: integer }
        predictedScoreB: { type: integer }
        predictedExtraTimeA: { type: ["integer", "null"] }
        predictedExtraTimeB: { type: ["integer", "null"] }
        predictedPenaltiesA: { type: ["integer", "null"] }
        predictedPenaltiesB: { type: ["integer", "null"] }
        points: { type: ["integer", "null"] }
        resultLabel:
          type: ["string", "null"]
          enum: ["Exact Score", "Goal Difference", "Correct Result", "One Team Score", "Missed", null]
        status: { type: string, enum: [submitted, locked, scored] }
        submittedAt: { type: string, format: date-time }
...
    CsvImportInput:
      type: object
      required: [csv]
      properties:
        csv: { type: string, minLength: 1 }
    ImportSummary:
      type: object
      required: [matchesCreated, teamsCreated, skippedDuplicates, errors]
      properties:
        matchesCreated: { type: integer }
        teamsCreated: { type: integer }
        skippedDuplicates: { type: integer }
        errors:
          type: array
          items: { type: string }
    LineupPlayer:
      type: object
      required: [name]
      properties:
        name: { type: string }
        position: { type: ["string", "null"] }
        shirtNumber: { type: ["integer", "null"] }
    TeamLineup:
      type: object
      required: [teamId, teamName, lineup, bench]
      properties:
        teamId: { type: integer }
        teamName: { type: string }
        formation: { type: ["string", "null"] }
        coach: { type: ["string", "null"] }
        lineup:
          type: array
          items: { $ref: "#/components/schemas/LineupPlayer" }
        bench:
          type: array
          items: { $ref: "#/components/schemas/LineupPlayer" }
    MatchLineups:
      type: object
      required: [matchId, provider, cachedAt, home, away]
      properties:
        matchId: { type: integer }
        provider: { type: string }
        cachedAt: { type: ["string", "null"], format: date-time }
        unavailableReason: { type: ["string", "null"] }
        home: { $ref: "#/components/schemas/TeamLineup" }
        away: { $ref: "#/components/schemas/TeamLineup" }
...